+++ forge/forge/op/eval/forge/__init__.py
@@ -78,8 +78,8 @@ op_to_module_map = {
     "reduce_avg": "reduce",
     "reduce_sum": "reduce",
     "reduce_max": "reduce",
-    "conv2d": "convolution",
-    "conv2d_transpose": "convolution",
+    "conv2d": "misc",
+    "conv2d_transpose": "misc",
     "conv3d": "convolution",
     "max_pool1d": "pooling",
     "max_pool2d": "pooling",
